Python/requirements.txt
pyqt5
pyserial==3.5
pyinstaller
pymodbus==2.5.3